requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
dill==0.3.6
filelock==3.9.1
Flask==2.2.3
flatbuffers==23.3.3
frozenlist==1.3.3
fsspec==2023.3.0
huggingface-hub==0.13.2
humanfriendly==10.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
numpy==1.24.2
onnxruntime==1.14.1
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==4.22.1
pyarrow==11.0.0
pyreadline3==3.4.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
responses==0.18.0
six==1.16.0
sympy==1.11.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
xxhash==3.2.0
yarl==1.8.2
zipp==3.15.0
